+++ frontend/src/app/_components/carousel/carousel.component.html
@@ -1,0 +1,31 @@
+<ngb-carousel #carousel class="carousel" *ngIf="imageBuffer.length > 0"
+              [showNavigationArrows]="true" [showNavigationIndicators]="false"
+              [pauseOnHover]="false" [pauseOnFocus]="false" [interval]="20000"
+              (slid)="handleSlidEvent($event)">
+  <ng-template ngbSlide *ngFor="let image of imageBuffer">
+    <div class="image-wrapper">
+      <img class="image" [src]="image.src" [alt]="'Picture of \'' + image.caption + '\''">
+    </div>
+    <div class="carousel-caption caption">
+      {{image.caption}}
+    </div>
+  </ng-template>
+</ngb-carousel>
+
+<!-- Gradient overlay -->
+<div class="overlay"></div>
+
+<!-- Pause Button -->
+<div class="playpause" (click)="pause()" *ngIf="!paused">
+  <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" class="bi bi-pause-circle" viewBox="0 0 16 16">
+    <path fill-rule="evenodd" d="M8 15A7 7 0 1 0 8 1a7 7 0 0 0 0 14zm0 1A8 8 0 1 0 8 0a8 8 0 0 0 0 16z"/>
+    <path fill-rule="evenodd" d="M5 6.25a1.25 1.25 0 1 1 2.5 0v3.5a1.25 1.25 0 1 1-2.5 0v-3.5zm3.5 0a1.25 1.25 0 1 1 2.5 0v3.5a1.25 1.25 0 1 1-2.5 0v-3.5z"/>
+  </svg>
+</div>
+<!-- Play Button -->
+<div class="playpause" (click)="play()" *ngIf="paused">
+  <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" class="bi bi-play-circle" viewBox="0 0 16 16">
+    <path fill-rule="evenodd" d="M8 15A7 7 0 1 0 8 1a7 7 0 0 0 0 14zm0 1A8 8 0 1 0 8 0a8 8 0 0 0 0 16z"/>
+    <path fill-rule="evenodd" d="M6.271 5.055a.5.5 0 0 1 .52.038l3.5 2.5a.5.5 0 0 1 0 .814l-3.5 2.5A.5.5 0 0 1 6 10.5v-5a.5.5 0 0 1 .271-.445z"/>
+  </svg>
+</div>
