+++ frontend/src/app/app.component.html
@@ -1,6 +1,5 @@
 <app-navigation></app-navigation>
 <div class="container">
   <router-outlet></router-outlet>
-  <h3>It should be fairly obvious this is a placeholder.</h3>
   <app-settings></app-settings>
 </div>
